Add files via upload
Artemis Financial, which specializes in financial consulting has hired Global Rain, and I have been asked to develop a web application to modernize their operations. Artemis Financial deals in plans for savings, retirement, investments, and insurance for their customers. Global Rain has been hired to develop an application for customers to upload files and access accounts, and their primary concern is the verification of data transfer. I did very well, generating checksum values based on different encryption algorithms to ensure data validation regardless of security protocol. Secure coding is the most important aspect of coding, regardless of the scope of the application. The goal is to ensure the customers' data security as well as that of the company is considered during every step of the Software Development Life Cycle. I had a challenging time comprehending the vulnerability report generated by using Maven dependency checks. It was really the only part I found to be especially difficult. There is information. I had to do research. Dependency checks are extremely important, especially for applications that rely heavily on frameworks and external APIs to gather and manipulate data. My approach to adding security layers to the application would be the use of multiple encryption algorithms. Distinct aspects of the application require different approaches to security based on the action being taken. For example, some algorithms are easier/better suited for validating the data within a file, while still others are more focused on user credential authentication. A multi-layered technique of encryption/security is a practice I am looking forward to in future applications and projects. I began evaluating each file individually before testing for proper integration of internal and external resources. Once changes to the code were made, the process of running additional vulnerability checks helped ensure the correct vulnerability was suppressed. To help assist in completing the report, documentation for the Spring framework, as well as YouTube tutorials on how to generate and read vulnerability reports were a large part of developing this project. The resources for a specific tool or programming language can be helpful when working with a tool with which you are unfamiliar. A notable example of my work is the vulnerability report. Having the knowledge and understanding of security reports are essential to web development. I am not completely confident in my abilities to read and interpret these reports, but I do have enough knowledge to help identify vulnerabilities which can be a valuable tool to have when working with a development team.